requirements.txt
# AI Traffic Management System - Backend Requirements
# ===================================================

# Core AI / Model
torch>=1.12.0
torchvision>=0.13.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.7.0
Pillow>=9.0.0

# Backend Web Server
Flask>=2.2.0
Flask-CORS>=3.0.10

# JWT Authentication
PyJWT>=2.6.0

# Utilities
numpy>=1.21.0
tqdm>=4.64.0

# =====================
# System Requirements
# =====================
# Python >= 3.8
#
# Installation:
# pip install -r requirements.txt
#
# Note:
# - Runs on CPU
# - No GPU required